refactor: move parseJSONAttributes
