+++ pkg/fs/file.go
@@ -1,10 +1,8 @@
 package fs
 
 import (
-	"fmt"
 	"io"
 	"iter"
-	"strings"
 	"time"
 
 	"github.com/maruel/natural"
@@ -28,39 +26,6 @@ const (
 	SortAsc SortOrder = iota
 	SortDesc
 )
-
-// JSONAttributes selects optional attributes written to an analysis export.
-// A nil set preserves the complete legacy export.
-type JSONAttributes map[string]struct{}
-
-// ParseJSONAttributes parses the comma-separated --output-attrs value.
-func ParseJSONAttributes(value string) (JSONAttributes, error) {
-	if value == "" {
-		return nil, nil
-	}
-
-	attributes := make(JSONAttributes)
-	for _, attribute := range strings.Split(value, ",") {
-		attribute = strings.TrimSpace(attribute)
-		switch attribute {
-		case "name", "asize", "dsize", "items", "mtime", "notreg":
-			attributes[attribute] = struct{}{}
-		default:
-			return nil, fmt.Errorf("unknown JSON output attribute %q", attribute)
-		}
-	}
-
-	return attributes, nil
-}
-
-// Includes reports whether an optional JSON attribute should be written.
-func (attributes JSONAttributes) Includes(attribute string) bool {
-	if attributes == nil {
-		return true
-	}
-	_, ok := attributes[attribute]
-	return ok
-}
 
 // Item is a FS item (file or dir)
 type Item interface {
